Actualización JPA Versión 1.0

+++ demo/src/main/java/com/example/demo/servicio/MascotaService.java
@@ -8,9 +8,9 @@ import com.example.demo.entidad.Mascota;
 
 public interface MascotaService {
 
-    public Mascota SearchById(int id);
+    public Mascota SearchById(Long id);
     public Collection<Mascota> SearchAll();
-    public void deleteById(int id);
+    public void deleteById(Long id);
     public void update(Mascota mascota);
 
     public void save(Mascota mascota);
